+++ car_information/car_information_service.py
@@ -1,10 +1,30 @@
 import requests
 from bs4 import BeautifulSoup
+import base64
+import json
+import os
 
 
+SECRET_KEY = 'sk_DEMODEMODEMODEMODEMODEMO'
+OPEN_ALPR_ENDPOINT = 'https://api.openalpr.com/v2/recognize_bytes' \
+    + '?recognize_vehicle=1&country=eu&secret_key=%s' % (SECRET_KEY)
 REGNR_ENDPOINT = "https://regnr.info/husker-ikke-hele-regnr?"
 VEGVESENET_API_ENDPOINT = "https://www.vegvesen.no/ws/no/vegvesen/kjoretoy/" \
     + "kjoretoyoppslag/v1/kjennemerkeoppslag/kjoretoy/"
+
+
+# NOTE: Only one licence plate should be present in each image.
+def identify_license_plate(path, image):
+    with open(path + image, 'rb') as file:
+        img_base64 = base64.b64encode(file.read())
+
+    try:
+        r = requests.post(OPEN_ALPR_ENDPOINT, data=img_base64)
+        result_json = r.json()
+
+        return result_json["results"][0]["plate"]
+    except:
+        return None
 
 
 def scrape_from_regnr(license_plate):
